+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add `resolve` subcommand to resolve identifiers to their canonical form without starting any containers
+
 ## [0.2.4] - 2026-05-01
 
 ### Fixed